Change values from float type to numeric

+++ migrations/20190905125543-create-transaction.js
@@ -9,7 +9,7 @@ module.exports = {
                 type: Sequelize.INTEGER
             },
             amount: {
-                type: Sequelize.FLOAT
+                type: Sequelize.NUMERIC(6,2)
             },
 
             description: {
