+++ models/payable.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
     const Payable = sequelize.define('Payable', {
         paymentDate: DataTypes.DATE,
         status: DataTypes.STRING,
-        amount: DataTypes.FLOAT,
+        amount: DataTypes.NUMERIC(6,2),
         transactionId: DataTypes.INTEGER
     }, {});
     Payable.associate = function(models) {
